***** RENAMED ALL HTML FILES TO .PHP *******
All HTML files needed to be changed .PHP in order to allow for embedded
PHP.
All pages now check if user/admin is logged in, if not then redirect to
Home.php.
If a user is logged in and tries to go to Home.php or Register.php, they
will be redirected to UserProfile.php.
If an admin is logged in and tries to go to Home.php or Register.php,
they will be redirected to AdminPage.php.
*** .php extension must be used in on HTML pages from now on